Include the latest message when watching
The function channel.getMessagesBefore(message) does not return the
message itself, only the ones before. This is the latest message so we
miss it.

But we also save it as last processed in the DB, so we won't read it
with the next message eiter, which means that we're effectively not
watching the channel.

The fix adds the requested message as first in the stream thus fixing
the problem.

Fixes issue #19.